+++ config/mssql-changes.php
@@ -2,4 +2,5 @@
 
 return [
     'connection' => env('MSSQL_CHANGES_CONNECTION', 'default'),
+    'columns-changed-max-width' => env('MSSQL_COLUMNS_CHANGED_MAX_WIDTH', 70),
 ];