+++ src/Actions/ShowChanges.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use Patabugen\MssqlChanges\Change;
 use Patabugen\MssqlChanges\Table;
+use Symfony\Component\Console\Helper\Table as ConsoleTable;
+use Symfony\Component\Console\Helper\TableSeparator;
 use Patabugen\MssqlChanges\Traits\HasVersionFiltersTrait;
 
 class ShowChanges extends BaseAction
@@ -50,12 +52,29 @@ class ShowChanges extends BaseAction
 
             return;
         }
-        $headers = array_keys($changes->first()->toArray());
-        $rows = $changes->map(function (Change $table) {
-            return $table->toArray();
-        })->toArray();
+        $table = new ConsoleTable($command->getOutput());
+        $table->setHeaders(array_keys($changes->first()->toArray()));
+        // Turn the changes into an array (and word-wrap the Columns list)
+        $changes = $changes->map(function (Change $tableChanges){
+            $row = $tableChanges->toArray();
+            $row['Columns Changed'] = wordwrap(
+                $row['Columns Changed'],
+                config('mssql-changes.columns-changed-max-width')
+            );
+            return $row;
+        });
+        // Add a separator between each row to make it easier to read
+        $rows = [];
+        foreach ($changes as $change) {
+            $rows[] = $change;
+            $rows[] = new TableSeparator;
+        }
+        // Remove the duplicate sepratator at the end
+        array_pop($rows);
 
-        $command->table($headers, $rows);
+        // Set the rows.
+        $table->setRows($rows);
+        $table->render();
         $command->info(
             count($changes).' '.Str::plural('change', $changes).' found'
         );